fix(panel): retire whole schema authority on class reads (#1709)
