+++ browser_tests/unit/add-node-command-budget.test.mjs
@@ -239,6 +239,7 @@ function realGraphAddNode({
     recordObjectInfoTypes: (defs) => defs,
     objectInfoHistory: { wasTypeEverDefined: () => false },
     objectInfoSnapshot: { record: () => true, clear: () => {} },
+    objectInfoCache: createObjectInfoCache(),
     verifiedNodeDefCache: createVerifiedNodeDefCache(),
     backendReconnectEpoch: 0,
     readPackImportFailures: async () => [],
